Phase 2- Created Enum class for the courses and a ProjectRandomGenerator class for chosing a random course for a trainee.
Phase 2- Created Enum class for the courses, and a ProjectRandomGenerator class for choosing a random course (safe) for a trainee.

In Trainee class- Created boolean isWaiting variable (can be used for the waiting list), assigned a random course method, also created get/set methods.

Tests for ProjectRandomGenerator and Courses- In process.